requirements.txt
torch
torchvision
Pillow
tqdm
pandas
matplotlib
icrawler
opencv-python
ultralytics
huggingface_hub
fastapi
uvicorn
pydantic
httpx
